+++ app/src/main/java/com/example/criteriaparser/model/DollarTwo.java
@@ -15,5 +15,5 @@ public class DollarTwo {
     public String type;
     @SerializedName("values")
     @Expose
-    public List<Integer> values = null;
+    public List<Double> values = null;
 }
